requirements.txt
aiogram>=3.4.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
